+++ src/routes/HistoryStack.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 
-import HomeScreen from "../screens/HomeScreen";
 import HistoryScreen from "../screens/HistoryScreen";
 import colors from "../config/colors";
 
@@ -19,11 +18,6 @@ const HistoryStack = () => {
         name="HistoryScreen"
         component={HistoryScreen}
         options={{ title: "History" }}
-      />
-      <Stack.Screen
-        name="HomeScreen"
-        component={HomeScreen}
-        options={{ title: "Home" }}
       />
     </Stack.Navigator>
   );
