+++ src/routes/not-using/CalendarStack.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 
-import HomeScreen from "../screens/HomeScreen";
-import CalendarScreen from "../screens/CalendarScreen";
-import colors from "../config/colors";
+import HomeScreen from "../../screens/HomeScreen";
+import CalendarScreen from "../../screens/CalendarScreen";
+import colors from "../../config/colors";
 
 const Stack = createStackNavigator();
